flask_api/requirements.txt
certifi==2020.12.5
cffi==1.14.0
chardet==3.0.4
click==7.1.1
defusedxml==0.6.0
Flask==1.1.2
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
numpy==1.18.2
pandas==1.0.3
pickleshare==0.7.5
protobuf==3.14.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
scikit-image==0.17.2
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
typing-extensions==3.7.4.3
urllib3==1.25.8
Werkzeug==1.0.1
